\name{compas1}
\alias{compas1}

\title{
  Criminal Offenders Screened in Florida
}

\description{ 
  Via FairML: A collection of criminal offenders screened in Florida (US) during 2013-14. This data was used to 
  predict recidivism (whether a criminal will reoffend or not) in the USA.
  
  Additional details for this dataset can be found via the fairML package. 
}
