\name{lak}
\alias{lak}

\title{
  Labor Market Discrimination
}

\description{
Fictional CVs sent to real employers to investigate discrimination via
given names. See Mullainathan and Bertran (2004).
}

\references{
   \itemize{
     \item Mullainathan, S. and Bertran, M. (2004). Are Emily and Greg More
     Employable Than Lakisha and Jamal? A Field Experiment on 
     Labor Market Discrimination. American Economic Review, 94:991-1013
   }
}

