% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{limitRange}
\alias{limitRange}
\title{Restricts the values of a data frame to specified limits}
\usage{
limitRange(data, yNames, xlim = NULL, ylim = NULL, zlim = NULL)
}
\arguments{
\item{data}{A data frame}

\item{yNames}{A preferably numeric vector that contains the indices/names 
of the columns by which the data frame is restricted}

\item{xlim}{A vector containing (min, max) to limit the 1st item of yNames to.}

\item{ylim}{A vector containing (min, max) to limit the 2nd item of yNames to.}

\item{zlim}{A vector containing (min, max) to limit the 3rd item of yNames to.}
}
\value{
A data frame that may or not be smaller than the original data frame
}
\description{
Restricts the values of a data frame to specified limits
}
\examples{
limitRange(svcensus, "wageinc", c(0, 200000))
}
