% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{makeSGroups}
\alias{makeSGroups}
\title{Generates a list of groups that exist within a sName column of a data frame}
\usage{
makeSGroups(data, sName, numGroups = NULL, sortedBy = "Name")
}
\arguments{
\item{data}{A dataframe}

\item{sName}{An index or name of a preferably factor column}

\item{numGroups}{A number of how many sGroups to generate}

\item{sortedBy}{A string indicating how the sGroups are generated
Possible values: 
  "Name" : first groups alphabetically are returned
  "Frequency" : most frequently occuring groups are returned
  "Frequency-Descending" : least frequently occuring groups are returned}
}
\value{
A vector of values that the sName of the data frame can take
}
\description{
Generates a list of groups that exist within a sName column of a data frame
}
\examples{
makeSGroups(svcensus, "educ")
}
