% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{makeSName}
\alias{makeSName}
\title{Generates an sName of a data frame}
\usage{
makeSName(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A string of the name of a factor column with the least amount of unique values
}
\description{
Generates an sName of a data frame
}
\examples{
makeSName(svcensus)
}
