% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{makeYNames}
\alias{makeYNames}
\title{Generates the yNames of a dataframe}
\usage{
makeYNames(data, count = 1)
}
\arguments{
\item{data}{A data frame}

\item{count}{A number of how many yNames to generate}
}
\value{
A numeric vector that corresponds to one or more numeric columns in the data frame
}
\description{
Generates the yNames of a dataframe
}
\examples{
makeYNames(svcensus, 3)
}
