\name{svcensus}
\alias{svcensus}

\title{
  Silicon Valley programmers and engineers data
}

\description{
This data set is adapted from the 2000 Census (5\% sample, person
records).  It is mainly restricted to programmers and engineers in the 
Silicon Valley area.  (Apparently due to errors, there are some from
other ZIP codes.)

There are columns for age, education, occupation, gender, wage income 
and number of weeks worked.  The education column has been collapsed to 
Master's degree, PhD and other. Income from 6 different occupations is available.

The variable codes, e.g. occupational codes, are available from 
\url{https://usa.ipums.org/usa/volii/occ2000.shtml}.
(Short code lists are given in the record layout, but longer ones are in
the appendix Code Lists.)
}
