\name{utilities}
\alias{getSuggestedLib}

\title{
Utitlities
}

\usage{
   getSuggestedLib(pkgName)
}

\arguments{
   \item{pkgName}{Name of the package to be checked/loaded.}
}
\description{

Attempts to load the specified package, halting execution upon failure.

}

\value{
No value, just side effects.
}

