\name{compas1}
\alias{compas1}

\title{
Criminal Offenders Screened in Florida
}

\description{ 

A collection of criminal offenders screened in Florida (US) during 
2013-14. This data was used to predict recidivism. 
  
Additional details for this dataset can be found via the \pkg{fairml} package. 

}

