\name{dsldConfounders}
\alias{dsldConfounders}
\title{dsldConfounders}

\description{
    WORK IN PROGRESS
}

\usage{
    dsldConfounders <- function(data, sName = NULL)
}

\arguments{
    \item{data}{
        dataframe with labeled column names
    }
    \item{yName}{
        as in qeML functions, predictive variable; should be a character or
        vector of column names
    }
    \item{sName}{
        name of the sensitive variable column, an R factor
    }
}

\author{
    T. Abdullah, Arjun A., Jonathan T.
}

\examples{
    
}

