\name{dsldDensityByS}
\alias{dsldDensityByS}
\title{Graphs densities of a response variable, grouped by a sensitive variable}
\usage{
dsldDensityByS(data, yName = NULL, sName = NULL, fill = FALSE)
}
\arguments{
\item{data}{A dataframe with 1 numerical column and a factor column}

\item{yName}{A name or index of the numerical column}

\item{sName}{A name or index of the factor column}

\item{fill}{A logical value determining if the graphed curve should be filled in}
}
\description{
Graphs densities of a response variable, grouped by a sensitive variable
}
