\name{dsldQeFairRF}
\alias{dsldQeFairRF}

\title{DSLDQeFairRF}

\description{ 
  To-do
}

\usage{
  To-do
}

\arguments{
  To-do
}

\author{
    N. Matloff
}

\examples{  
  # To-do
  library(dsld)
  data("svcensus")
  z <- dsldQeFairRF(data=svcensus,yName='wageinc',deweightPars=list(occ=0.2),sensNames='gender')
  z$testAcc
}

\details{
    To-do
}
        
