\name{dsldQeFairRidgeLin}
\alias{dsldQeFairRidgeLin}

\title{DSLDQeFairRidgeLin}

\description{ 
  To-do
}

\usage{
  To-do
}

\arguments{
  To-do
}

\author{
    N. Matloff
}

\examples{  
  # To-do
  library(dsld)
  data("svcensus")
  z <- dsldQeFairRidgeLin(data=svcensus,yName='wageinc',deweightPars=list(occ=0.2),sensNames='gender')
  z$testAcc
}

\details{
    To-do
}
        
