\name{dsldScatterPlot3D}
\alias{dsldScatterPlot3D}
\title{ScatterPlot3D in dsld}
\description{ The function, dlsdScatterPlot3D, creates a 3D scatterplot
that paints each unique group (a set of points) as a different color. These
groups are plotted against the 3 axes. For example: Looking at a dataset
that contains the unique groups catchers and shortstop. We can plot these two
groups against the 3 axes, age, height, weight. The function will produce a 3D
scatterplot of this data.}
\usage{dsldScatterPlot3D(data, grpcol=NULL, axiscols=NULL, grpnames=NULL, angle=40,  
                               sortedby="Frequency", numgrps=3, 
                               colors=NULL, pchs=NULL, main=NULL, sub=NULL, 
                               xlim=NULL, ylim=NULL, zlim=NULL)}

\arguments{
  \item{data}{
  a dataframe that has at least 3 numerical columns for the 3 axes.
  }
  \item{grpcol}{
  the index or name of the column that contains the groups for which the data
  will be grouped by. This column should not be numeric. 
  If not supplied, the function will use a non-logical column with the least 
  number of unique values.
  }
  \item{axiscols}{
  a vector of the indices or names of the columns that are graphed on the 
  3 axes. These columns should have numeric data.
  If not supplied, the function will use the first 3 numeric columns in the dataframe.
  }
  \item{grpnames}{
  a vector of the names of the groups for which the data will be grouped by. 
  Every value in the vector must exist in the grpcol column of the dataframe.
  If not supplied, the function will create it automatically according to the
  sortedby and numgrps parameters.
  By default, the function uses the 3 most frequently occuring values of the gpcol.
  }
  \item{angle}{
  the view angle of the scatterplot, in degrees.
  }
  \item{sortedby}{
  controls how grpnames is created automatically. If grpnames is suppled, this 
  does nothing. One of three values: "Name", "Frequency", "Frequency-Descending". 
  "Name" gets the first values alphabetically.
  "Frequency" gets the most frequently occuring values.
  "Frequency-Descending" gets the least frequently occuring values.
  }
  \item{numgrps}{
  how many groups will be automatically generated by the function. If grpnames
  is supplied, this does nothing. 
  }
  \item{colors}{
  a vector of colors that will correspond to each group on the graph. 
  Can also be a single value that will be repeated.
  By default, the function uses the numbers 2:numgrps+1 as the color vector
  }
  \item{pchs}{
  a vector of symbols that will correspond to each group on the graph.
  Can also be a single value that will be repeated.
  By default, the function uses the numbers 1:numgrps as the symbol vector.
  }
  \item{main}{
  the title of the graph. By default, its the name of the grpcol "vs. " the names of the axiscols
  }
  \item{sub}{
  the subtitle of the graph. Nothing by default.
  }
  \item{xlim, ylim, zlim}{
  the x, y and z limits (min, max) of the plot.
  }
}

\value{
  A visual of the 3D scatterPlot
}
\references{
Scaterrplot3D function:
\link{ https://cran.r-project.org/web/packages/scatterplot3d/index.html}
}
\author{
Jonathan Tran and Brandon Zarate }

\examples{
# Example 1: Using dataframe alone
library(qeML)
data(mlb)
dsldScatterPlot3D(mlb, main = "Example 1")

# Example 2: Using column names
data(pef)
axisColumns <- c("educ", "wageinc", "occ")
groupColumn <- "sex"

dsldScatterPlot3D(pef, groupColumn, axisColumns, main= "Example 2")

}

