\name{dsldScatterPlot3D}
\alias{dsldScatterPlot3D}
\title{ScatterPlot3D in dsld}
\description{ Plots a dataset on 3 axes, with the color of the point depending 
on a 4th variable.}
\usage{
dsldScatterPlot3D(data, yNames, sName, sGroups = NULL, sortedBy =
  "Name", numGroups = 8, maxPoints = NULL, xlim = NULL,
  ylim = NULL, zlim = NULL, main = NULL, colors =
  "Paired", opacity = 1, pointSize = 8)}

\arguments{
  \item{data}{
  A dataframe with at least 4 columns as response variables to be graphed on the 
  3 axes, with an additional columm may be used as a sensitive variable.
  }
  \item{yNames}{
  A vector of the indices or names of the columns of the dataframe to be 
  graphed on the 3 axes. 
  }
  \item{sName}{
  The index or name of the column that contains the groups for which the data
  will be grouped by. This will affect the colors of the points of the graph. 
  This column must not be numeric. 
  }
  \item{sGroups}{
  A vector of the names of the groups for which the data will be grouped by. 
  Every value in the vector must exist in the sName column of the dataframe.
  If not supplied or is NULL, the function will create this automatically 
  according to the sortedby and numgrps parameters.
  By default, the function uses the 8 alphabetically first distinct groups 
  in the sName column.
  }
  \item{sortedBy}{
  Controls how sGroups is created automatically. If sGroups is suppled, this 
  does nothing. One of three values: "Name", "Frequency", "Frequency-Descending". 
  "Name" gets the first values alphabetically.
  "Frequency" gets the most frequently occuring values.
  "Frequency-Descending" gets the least frequently occuring values.
  }
  \item{numGroups}{
  How many groups will be automatically generated by the function. If grpnames
  is supplied, this does nothing. 
  }
  \item{maxPoints}{
  A limit to how many points may be displayed on the graph.
  There is no limit by default.
  }
  \item{xlim, ylim, zlim}{
  The x, y and z limits- a vector with c(min, max)- the plot will 
  graph according to.
  }
  \item{main}{
  The title of the graph. By default, its the name of the sName "vs. " the yNames
  }
  \item{colors}{
  Either a colorbrewer2.org palette name (e.g. "YlOrRd" or "Blues"), 
  or a vector of colors to interpolate in hexadecimal   "#RRGGBB" format, 
  or a color interpolation function like colorRamp().
  }
  \item{opacity}{
  A value between 0 and 1.
  }
  \item{pointSize}{
  A value above 1.
  }
}

\references{
   https://plotly.com/r/3d-scatter-plots/
}

\author{
J. Tran and B. Zarate
}

\examples{
# Example 1: Using sName, yNames, and xlim
data("law.school.admissions")
dsldScatterPlot3D(law.school.admissions, sName = "race1", 
  yNames=c("ugpa", "lsat","age"), xlim=c(2,4))
}

