\name{dsldScatterPlot3D}
\alias{dsldScatterPlot3D}
\title{ScatterPlot3D in dsld}
\description{ Plots a dataset on 3 axes, with the color of the point depending 
on a 4th variable.}
\usage{dsldScatterPlot3D(data, sName=NULL, yNames=NULL, sGroups=NULL, 
                              sortedBy="Name", numGroups=8, maxPoints=NULL, 
                              main=NULL, colors="Paired", opacity=1, 
                              pointSize=8)}

\arguments{
  \item{data}{
  a dataframe with at least 3 columns for the axes. Additional columms may be
  used as a sensitive variable.
  }
  \item{sName}{
  the index or name of the column that contains the groups for which the data
  will be grouped by. This column should not be numeric. 
  If not supplied, the function will use a non-logical column with the least 
  number of unique values.
  }
  \item{yName}{
  a vector of the indices or names of the columns that are graphed on the 
  3 axes. These columns should have numeric data.
  If not supplied, the function will use the first 3 numeric columns in the dataframe.
  }
  \item{sGroups}{
  a vector of the names of the groups for which the data will be grouped by. 
  Every value in the vector must exist in the sName column of the dataframe.
  If not supplied, the function will create it automatically according to the
  sortedby and numgrps parameters.
  By default, the function uses the 8 alphabetically first distinct groups 
  in the sName column.
  }
  \item{sortedBy}{
  controls how sGroups is created automatically. If sGroups is suppled, this 
  does nothing. One of three values: "Name", "Frequency", "Frequency-Descending". 
  "Name" gets the first values alphabetically.
  "Frequency" gets the most frequently occuring values.
  "Frequency-Descending" gets the least frequently occuring values.
  }
  \item{numGroups}{
  how many groups will be automatically generated by the function. If grpnames
  is supplied, this does nothing. 
  }
  \item{maxPoints}{
  a limit to how many points may be displayed on the graph.
  There is no limit by default.
  }
  \item{main}{
  the title of the graph. By default, its the name of the sName "vs. " the yNames
  }
  \item{colors}{
  either a colorbrewer2.org palette name (e.g. "YlOrRd" or "Blues"), 
  or a vector of colors to interpolate in hexadecimal   "#RRGGBB" format, 
  or a color interpolation function like colorRamp().
  }
  \item{opacity}{
  a value between 0 and 1.
  }
  \item{pointSize}{
  a value above 1.
  }
}

\value{
  An interactable 3D scatterPlot
}
\references{
plotly's 3 scatterplot:
\link{ https://plotly.com/r/3d-scatter-plots/}
}
\author{
Jonathan Tran and Brandon Zarate }

\examples{
# Example 1: Using dataframe alone
library(qeML)
data(mlb)
dsldScatterPlot3D(mlb, main = "Example 1")

# Example 2: Using column names
data(pef)
sName <- "gender"
yNames <- c("educ", "wageinc", "occ")

dsldScatterPlot3D(pef, sName = sName, yNames = yNames, main = "Example 2", 
pointSize = 5)

}

