\name{mortgageSE}
\alias{mortgageSE}

\title{
  Mortgage Denial
}

\description{

The dataset provides applicant information (including race, income, loan
information, etc.) The response variable indicates whether or not the
applicant was approved for the loan. Additional details can be found in
the \code{SortedEffects} package.

}
