\name{svcensus}
\alias{svcensus}

\title{
  Silicon Valley programmers and engineers data
}

\description{
  Via qeML: This data set is adapted from the 2000 Census (5% sample, person records). 
  It is mainly restricted to programmers and engineers in the Silicon Valley area. 

  Additional details for this dataset can be found via the qeML package. 
}
